Reworking options classes
Now there is a base class that handles the reading of an options file, and
children classes (only one at the moment) that take care of actually mapping
the option values to variables in memory. Children classes default their values
in the constructor initialization list, and then load the values available from
the option collection in the constructor body.

The option files now should include option groups (lines that look like
"[group]") to have nicer namespaces for option names.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>