No need to pass down so many shared_ptr
Since we are not changing here (yet) how objects are bound to each other
there's no need to pass down pointers. We thus de-reference the pointers early
on (i.e., when still on evolve_galaxies), and pass down simply the references
through all the internal methods that read/write them.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>