AGN feedback model implemented. ODEs of SF updated.
Two models for AGN feedback are implemented: the galform model (as
described in Lacey+16), and the L-galaxies models, as described in
Croton+16. Those models modify the cooling rates if AGN feedback is on,
and that is done in gas_cooling. The ODEs in physical_model.cpp are also
updated to comply with halo having three gas components: cold and hot
gas, and ejected gas.