*Parameters classes do not inherit from Options
This inheritance was causing the configuration file to be read more than
once unnecessarily. Now the *Parameters classes receive the fully-parsed
Options object instead as a reference, which then they use to load the
specific configuration values into their members.

Because of this change, the is_skipable method from Options has now been
moved to the utils.cpp file for better reuse.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>