Galaxy mergers and starburst routines finalized.
In galay merger routines I now included the part where I classify the
merger and drive the starbust in the cases where this is required. For
that it was also necessary to modify the physical model to include
routines that do the starburst. 

I also included several functions in the galaxy class that calculate
some useful properties, such as disk, bulge and baryon mass.