Definition of central subhalos done self-consistently
Now central subhalos are define within SHARK in a self-consistent way
following the ascendants of the most massive subhalos. In addition
accreted mass is calculated there so that we can then calculate the
accreted baryon mass in each snasphot. 

I also commented the calculation of the orbital part of the dynamical
friction timescale because I need to check that the integration is being
done properly.