Less std::pow abuse
This is by no means extensive, as there are probably more portions of
the code that still invoke pow with known exponents (i.e., they should
be replaced by sqrt or cbrt), but sets a precedent.

Also, this fixes a problem where we were getting ones due to the
exponent being zero (which, in turn, happened because it was expressed
as a fraction of two integer numbers, the nominator being less than the
denominator).

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>