New GSL exception, thrown by new handler
Instead of aborting the program, we install our own GSL error handler,
which takes the error information and throws an exception with it. This
way we have more control over what to do in case of an error (by default
we are still printing it and finishing the program, but we might want to
do other things in the future).

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>