Recycling parameters read and bugs in physical model fixed
The recycling parameters weren't being read so now I implemented the
class to read those parameters. There were also some bugs in the
physical model, for example, the order of inputs for the star formation
rate was wrong, and the inputs for the stellar feedback calculation were
wrong.

I also implemented a new function to calculate the dark matter accreted
mass from snapshot to snapshot in the tree_builder.