Bugs fixed in galaxy mergers and in gas_cooling
There were some bugs in galaxy mergers as galaxies were being removed in
a for loop over galaxies themselves. This is now fixed. In gas_cooling
we are now using the spline2d interpolator to avoid giving the function
the tables of temperature, metallicity and lambda.

I also just started writing the writing routine, but needs development.