Correctly remove satellite galaxies
The previous incarnation of the code was simply wrong, as the
erase(first, last) signature works with iterators from the same
container. This new version should be correct (although I haven't
checked it). Also, we might want to think whether we want to change the
vector<GalaxyPtr> in the Subhalo class to a set, and have the central
galaxy stored separately for efficiency.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>