Several bugs fixed in stellar and agn feedback, and physical model
Metallicities of the cold and hot gas were wrong in physical_model and
are now correctly calculated. In addition, several bugs were fixed
related to the calculation of the outflow rate (beta parameter), and the
AGN feedback (e.g. seeding of halos with black holes, density_shell). I
also added a pre_enrichment metallicity which is a floor applied to the
metallicity and hot halos are first seeded with that metallicity when
first galaxies are created.