Function transfer_galaxies and galaxy_mergers modified
(1) Function transfer_galaxies: before this was done within the for loop
of halos and now it's done at the end of the snapshot. Within that
function we also added a counter of how many subhalos do not have
descendants and issue a warning in those cases.

(2) Galaxy mergers: added SMBH growth during starbursts. There was also
a bug in the case of minor mergers that trigger a starburst, because in
that case the disk gas mass should be transferred to the bulge and that
wasn't happening. This is now fixed.