Bug in reincorporation mass corrected and metallicity floor applied.
There was a bug in the calculation of reincorporated mass because in the
case it was 0, the metallicity was giving me -nan. I also added a floor
to the metallicity when solving the equations of star_formation. I have
also for now added a routine that writes outputs to ascii.