Force ODE to return a solution even if it isn't converged
Force ODE to send a warning i it didn't converge but continue running.
ODE integrator in a now rkck (which is what I used in galform). Accuracy
of star formation is set to 5% as I had in galform. 

It seems that there are issues with the types of galaxies because there
are many CENTRAL in the same halo, which is obviously an issue. Check
how type 2 are assign and make sure halos have only one central subhalo,
and that central subhalo has only one central galaxy.