DarkMatterHalos class(es) redesigned
This is not a major redesign, but makes for easier-to-read code. Instead
of having several if statements depending on the dark matter profile
being used, we now have two new subclasses that derive from
DarkMatterHalos implementing the two (now virtual) methods that are
profile-dependant.

I'm not sure if this is the best way of separating this or not -- maybe
I should have created a new DarkMatterProfile class with the two virtual
methods and have it as a member of DarkMatterHalos, but I'm not sure
what DarkMatterHalos actually represents (a set of laws, a set of
equations related to dark matter halos, etc...?).

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>