numerical_constants more C++11 compliant
There were a few values in numerical_constants that used std::log and
std::exp, even though both of them are non-constexpr (but our variables
are). gcc was still happy with this, but clang was more pedantic and
didn't allow for it.

I've removed the usage of std::log (we already had two other log values
that could be used to calculate this one), and created a constexpr pow
function that works with (positive) integer exponents only, and that can
be used to define constants that use powers.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>