Manual integrator for SFR, and bugs in AGN feedback fixed.
Tests done in SFR intgrator. I implemented a manual integreation to
study so far how the z=0 galaxy properties are coming about. Several
bugs were found in the AGN feedback routine when the L-GALAXIES model
was used and are now fixed. In addition I changed a bit the routine that
calculates the merging timescales to include the different baryon masses
of the satellite galaxies. 

I also put back the ascii output for faster testing.