Different native/file HDF5 types
When writing data, we want to potentially use different datatypes to
describe the format of the data in memory, and the format of data in
disk. The hdf5 libs should automatically convert from one to the other
during the writing process, if required.

We don't need to specify a native datatype (or should it be called
"memory" datatype instead?) for the std::string and bool types, as we
overload already the dataset writing templated methods for these types.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>