Re-implementing option reading, allow ranges
This change re-implements the reading of values from the options file.
It centralizes the conversion from string to different built-in
datatypes, and also implements (in a common code) support for supporting
range specifications (available only for integer types), which can end
up being stored either as vectors or as sets.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>