Centralizing output directory creation
The logic to create the output directories structure was, so far,
centralized in the HDF5 Writer class, and therefore the ASCII writer was
(silently) failing to create its file if the directories didn't exist
beforehand.

This patch solves this, both ensuring that the output directories exist
for both formats, and also standarizing the output directories structure
itself.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>