Fixing definition of central Subhalos
Central subhalos (respective to their host Halo) are defined, in the
general case, by looking at snapshot i's central subhalos, finding their
most massive ascendant subhalo, and setting *that* subhalo as the
central subhalo of *its* hosting halo.

In the first iteration of this process (i.e., when we are looking at
Halos and Subhalos without descendants) we simply take on each halo the
most massive subhalo on each halo and set it as the central subhalo.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>