Redefinition of central subhalo and main progenitor
Main progenitor information is now read from the dhalos trees and is
used to define central subhalos (following the main progenitors of the
most massive subhalo at the lowest redshift). Because of these changes
the function define_main_progenitor is not needed anymore.

New function ensure_halo_mass_growth applies a strict mass growth
criterion for halos (halos have to be at least as massive as their main
progenitors). 

Function merging_subhalos has also been modified to deal with central
subhalos that disappear on the next snapshot. In order to accommodate
these cases better, the call of this function in the main() has been
moved after the populate_halos function. 

There are some problems remaining with the routine of merging_subhalos
apparently because of the part that deals with those central_subhalos
disappearing. Code currently crashes in merging_galaxies because it
doesn't find a central_galaxy in the central_subhalo.