Avoid copies of Halo in ascendants and merger tree
When linking Subhalos to establish their parentage relationship, we also
relate their corresponding Halos. However, due to the multiplicity
between Subhalos and Halos (one Halo contains zero or more Subhalos), we
were running the same relationship establishment logic for two Halos more than once.

This situation happened if two Subhalos SH1 and SH2, living in Halo H1,
were deemed as ascendants of Subhalos SH3 and SH4 respectively, both of
which live in Halo H2. In this case, when establishing
SH1-is-parent-of-SH3 we also established H1-is-parent-of-H2, and
repeated this operation when establishing SH2-is-parent-of-SH4, leading
finally to repeated references to H1 in the merger tree and in the list
of ascendants of H2.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>