Total baryons being tracked and written in the outputs
New class of total baryons to track the evolution of all the mass
components to make diagnostics easier. This required a new function in
evolve_halos to make the tracking, to move the function
get_molecular_gas to star formation so that it can be called by any
other function, and galaxy_writer now writes a new group in the HDF5
called Global/ with all these properties.