New calculation that keeps track of all baryons created
New variable in TotalBaryon class that keeps tracks of all baryons ever
created, which come from two sources: (i) first time a halo appears in
the simulation a fraction of the dark matter mass is assigned to the hot
halo gas, (ii) dark matter accretion brings additional baryons with the
same proportionality factor as applied in (i). That factor is set by the
cosmology. This was done to ensure mass conservation when keeping track
of all baryons in different reservoirs.