Stellar feedback model modified to make it general
- The new feedback model allows for two mass loadings: one from the
galaxy and another one from the halo. The gas that does not escape the
halo is reincorporated immediately to the hot gas in the set of ODEs in
the physical_model.

- Another important change is that we remove all subhalos that do not
have an idenditified descendant ID in the tree_builder.

- Gas cooling: the ejected reservoir of satellites is transfered to the
ejected reservoir of the central (rather than the hot gas reservoir).