Adding support for commentin datasets
Datasets have a special comment field that can be set to the value given
by the user. This field is not shown by hdfview however (but h5dump
will show it), and therefore people seem to have resorted to add a
"comment" attribute to the dataset, which does show up in hdf5view. We
are no less.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>