star formation history modified to reflect what ends up in a disk/bulge
What is needed to create the SED is to have separated the stellar mass
growth of what ends up in a disk and a bulge. Thus the code has been
modified to transfer the stellar growth histories from satellites to
centrals in the case of mergers and disks to bulges in the case of disk
instabilities. Code still seems to have some bug as in the output there
are discontinuities in the stellar mass growth history.