Using references instead of copies where possible
This is specially important in the galaxy_writer module, where the
TotalBaryon object (which internally holds many vectors with the history
of different properties) was being passed down and copied over.

The tree_builder also had a few of these, but the mostly had pointers
inside so didn't hurt us much.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>