Initial version of HPC scripts
These scripts are meant to be used within HPC facilities to submit and
run shark. They parallelize shark by spawning different shark instances
across the cluster, each processing a different subvolume of the input
data.

The scripts have been designed to be cluster-independent, although there
are still some minor bits in there that are still specific to Magnus (in
the Pawsey supercomputing centre, in Perth, Australia) and that need to
be cleaned up and generalized. They are also meant to support more than
one queueing system. At the moment they support SLURM only, but adding
support for other systems should not be that complex.

These scripts are installed together with the shark binaries and
library, making them easy for users to invoke them.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>