Quicker memory release after reading subhalo data
Previously we held onto the raw data read from the HDF5 files until all
the Halos were constructed from the Subhalos, which in turn are
constructed from the raw data. However, there is no need to hold onto
the raw data after the Subhalos are created, so we are now quickly
releasing it by having it live in a separate method that returns the
Subhalos read from the file *before* constructing the Halos.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>