Release list of Halos after creating MergerTrees
The list of Halos read from the SURFSReader object doesn't need to stay
in memory after we create the MergerTrees, because the trees already
contain the Halos, and we actually always looping through the structures
using the trees as the root of everything. We were previously not
clearing this vector, so at the end shark ended up using more memory
than necessary.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>