Fixing printing of _fixed structs
Because the _fixed structure is defined within the detail namespace, we
need to overload the basic_ostream::operator<< method within the same
namespace. Failing to do so means that the compiler will not be able to
find the overloading when trying to dump a _fixed object into an object
that is not part of the std namespace (like a boost logger, for
example).

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>