Initial OpenMP parallelization
The easiest way to parallelize shark is to process merger trees in
parallel. Merger trees are (currently) fully independent of each other,
so in principle they can be processed in parallel without any side
effects.

The number of threads is a user-provided parameters. This makes the
number of threads more explicit in the code, instead of relying in the
OMP_NUM_THREADS environment variables or other implicit ways of defining
the number of threads to use. However, we still give users the ability
to use whatever default the OpenMP implementation gives by specifying
"-t 0".

To properly run things in parallel we need to make sure we don't share
state across threads. The main stateful objects at the moment are the
GSL ODE system and the GSL integrator, and any class that directly and
indirectly uses them. We thus are now creating multiple copies of all
these objects as necessary. When the code is compiled without OpenMP
support a single copy of these objects is created, but the overall code
is still generic and assumes there is always a collection of such
objects.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>