Delaying the transfer of galaxies across snapshots
Galaxies were being transferred from a Subhalo to its descendant Subhalo
in two different places in the code. The first location was at the end
of the per-halo evolution loop via galaxy_mergers.merging_subhalos, and
the second was at the end of the per-snapshot evolution loop via
transfer_galaxies_to_next_snapshot. In between (i.e., after the per-halo
loop finished, but before the per-snapshot loop iteration finished) the
track_total_baryons function was assuming all galaxies were still living
in the current snapshot, and therefore it failed to create a history
item for those that had been transferred to the next snapshot already.

This change delayes the transfer *in the time dimension* of those
galaxies so it always happens at the end of iteration of the
per-snapshot evolution loop. This ensures that the total baryon mass
tracking, and other bits of logic that might depend on snapshot-local
data, can reach all galaxies living in the current snapshot before they
are transferred to the next.

Note that the galaxy_mergers.merging_subhalos function still performs
galaxy transfers when merging subhalos, but this is across subhalos that
are part of the same halo, and therefore they do not move across
snapshots.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>