New plotting scripts
These are the standard plotting scripts used to make sure the outputs
from shark actually make sense. This is the initial set, but more might
come in the future.

These scripts use the same configuration file used to run shark to find
out what simulation/model/shark directory should be used for
input/output, although these options can be explicitly specified as well
if needed (for more manual processing needs). All these options are
given as command-line parameters by the user, and therefore it is easy
to verify if the user is still missing some information or not.

Almost the scripts are quite flexible, some options still remain
hard-coded. These are only a few though, so it can be changed in the
future with relative ease.

Finally, scripts work both with python 2.7 and 3.