specific AM calculated; bug in vbulge fixed; velocity-mass plots
- here was a bug in the calculation of the circular velocity
contribution from bulges, which is now fixed.
- calculation of angular momentum of disks and bulges has now been
checked is working well.
- new option in stellar feedback galaxy_scaling which is set to true if
the user wants to use the velocity of the disk/bulge rather than that of
the subhalo for the mass loading calculation. This means that the galaxy
velocity of the disk or the bulge are given as parameters to the
physical model to allow the user more flexibility when calculating the
stellar feedback mass loading.

- new plots showing the disk and bulge velocities vs. stellar mass to
verify that their  values are sensible. Standard shark model produces
reasonable velocity-mass scalings.