new reionisation model, bugs on smhm.py fixed
- new reionisation model added. Now there are two options, the GALFORM
and the Sobacchi model. This meant to add a new class to reionisation.h
that contains the models. Parameters were also expanded to include the
model name and the power-law index of the redshift dependence in the
case of Sobacchi. 

- changes in main to reflect changes in reionisation. Now gas_cooling
receives reionisation as input rather than the reio_params. 

- SF test changes slightly to catch cases where SFR = 0 and Sigma_gas>0
and fmol > 0.