Changes to appropriately account for angular momentum evolution.
Several changes have been included to appropriately account for the
angular momentum evolution in galaxy components gas and disk:

- new parameters included to allow for angular momentum internal
calculation within disks and bulges.

- ODEs in physical model have been extended to include the equations
that follow the angular momentum exchange between components.

- Whether we assume stars form from the same sAM of the gas, or we allow
for the molecular gas to pass that information into the stellar
component, is encoded in star_formation*

- We now follow the specific angular momentum of the hot and cold halo
components.

- New functions of angular momentum of the disk, bulge and halo were
included in dark_matter_halos. All of the angular momentum functions are
within dark_matter_halos and star_formation.

- Galaxy mergers, gas cooling and disk instabilities were modified
accordingly to take into account these new functions. 

- Stellar feedback was also modified to allow for possible outflows with
preferential angular momentum losses. The standard shark model works
assuming that outflows carry out the same specific angular momentum than
the gas in the disk-bulge, but can could be modified in the future.

- There are still problems with (i) the bulge sizes and how to calculate
them during minor mergers, (ii) how I'm calculating jrate in the case
where jgas is assumed to be constant, as this should return a jgas after
the solution of the ODEs that is the same as when it started, but for
some reason that is not happening. This needs to be solved.