*Parameter classes default member initialization
This "new" (since C++11) way of initializing members is much more comfortable
to read than the usual constructor initialization list, and fits perfectly for
the *Parameters classes, which all define default values for their members.
This cuts the number of lines of code by quite a bit, while maintaining
functionality, decreasing possibility of wrong initialization order (a usual
source of compiler warnings), and increasing readibility, so it's a win for
everyone.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>