Sizes of disks self-consistently calculated now.
Modifications to the calculation of the angular momentum of disks were
introduced to allow for a self-consistent development of the specific
AM. With this in mind the part where AM is transfered to the disk/bulge
from the cooling gas is in the ODE system and AM does not enter anywhere
else.

- Bulge sizes are still not fully working and this is because bulges
need to be treated as pseudo angular momentum components (because of
their large dispersion contribution). This should be done as in Cole et
al. (2000).