Several changes done to remove cases where rscale is not well defined
- there was an outdated function, transfer_bulge_am, that was still
being called but it should not have been called. This happened in both
galaxy_mergers and disk_instabilities, so the calls were removed.

- In addition several checks were put in place so the code crushes if
rscale or sAM are not well defined.

- Changes in galaxy_write to add more info regarding simulation box and
runs.

- Changes in galaxy_writer and specifically in SFH to save only galaxies
that at z=0 have a stellar mass >0, and only the relevant properties as
needed by ProSpect. This required to add cosmology as an input to
galaxy_writer as ProSpect needs the look back times of the snapshots.

- As a result I removed the gas properties being tracked in the history
items as they are not needed by ProSpect.

- New property angular_momentum is included in the galaxy component.

- Also in galaxy_writer, I'm not outputting the angular momentum vector
of the galaxies type 1 and 0.