Solved bug in F(c) calculation and in sfr_z_disk/bulge
- There were a couple of bugs on the way f_c was calculated in
dark_matter_halos (that goes in the calculation of the velocity
dispersion of the halo). And another bug in the sfr_z_disk and bulge, as
they needed to be set to 0 at the end of each snapshot to avoid
accummulating them. I also changed the names of the entries in the hdf5
files of the star_formation_histories to make the naming more obvious.