get_molecular_gas returns molecular_gas
This method accepted several parameters that were given by reference
(and previously by pointer) in order to retrieve their values later.
This made it hard to maintain and use in a number of places. Now the
code is much easier to read and use. I've also renamed the jmol and
jatom variables into j_mol and j_atom for consistency.

Moreover, now that we want to try and keep these values it will be
easier to maintain one container of molecular_gas rather than many
containers for each individual quantity.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>