Reusing per-galaxy molecular_gas results
Instead of having the total baryon tracking and the galaxy writing
processes calculating both the molecular gas results for each galaxy we
calculate it once only and then reuse the results in both processes.
This saves us quite a lot of calculations which were otherwise
duplicated when writing galaxies onto disk.

Note that we need to calculate j_mol and j_atom only if we are meaning
to write the galaxies to disk, which was probably the original cause of
this duplication. We are preserving this behavior after this
refactoring, so we shouldn't incur into performance penalties even when
not writing galaxies.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>