Don't consider Subhalos we're not going to use
The simulation.min_snapshot preference gives a lower limit on the
snapshot on which we operate later on, so there is no point on even
creating Subhalos belonging to those snapshots in the first place.

This is the most efficient place to do this filtering, since will save
us memory on all later stages of the program.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>