Parallelising calculation of molecular gas
The usual drill takes place: per-thread copies of the required objects
are created, and aggregation happens at the end of the parallel block.
In this case we need to keep track of per-thread molgas_per_galaxy
objects, plus a StarFormation object for each thread to avoid shared
state.

This is the final major parallelisation point we needed to address
during the main evolution loop; after this the time spent in parallel
sections accounts for ~90% of the wall time in common cases measured so
far. After, parallelising more of the initial data loading and
verification procedures is the next main goal, but we are already in a
pretty good state.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>