Check redshift data is correct
We have come across redshift files that list the same redshift for
different consecutive snapshots. This caused errors down the line,
specifically because the delta_t for those snapshots was zero, which did
not play well with the GSL ODE solver of course. Depending on which
snapshots are affected, this error can appear either very quickly or
rather late during the execution of shark, which can be frustrating
during long runs.

This additional check gives a clear error message regarding the cause of
the failure, and makes shark fail quickly at startup instead of doing it
later during the main execution of the program.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>