Defining main progenitor when none is found
We use the most massive ascendant as the main progenitor when none of
them are marked as such. This shouldn't happen in the first place, but
instead of failing like we were doing before we are issuing a warning to
let users know that there is something wrong with the data (but let them
continue anyway).

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>