Moving most of main logic to shark_runner
The new SharkRunner class now contains all the high-level logic of a
shark instance: it creates all the individual objects needed throughout
the execution of shark, reads the initial data, creates the merger
trees, and runs the evolution of galaxies across the snapshots. This
makes the main module much slimmer and easy to handle. This is also a
better starting point to start breaking down the run() method, which
needs to be broken down for better error handling and code reading.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>