New CosmologyPtr type
This hides the fact that (currently) we serve cosmology pointers via
std::shared_ptr. A new make_cosmology templated function also hides the
creation of the Cosmology in particular (of which we currently have only
one though).

The constructor of the Cosmology class takes a const reference to the
parameters now too, since it was unnecessarily copying the parameters
object before.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>