New DarkMatterHalosPtr type
This hides the fact that (currently) we serve DarkMatterHalos pointers
through std::shared_ptr. A new make_dark_matter_halos template function
creates the appropriate underlying class depending on the
DarkMatterHalosParameters object passed down.

The constructor oft he DarkMatterHalos base class was also adjusted for
correctness, as the parameters object doesn't need to be constant.

I've also re-ordered the includes to be more consistent to how we are
doing them in the rest of the code.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>