Simplifying SharkRunner::run() method
Instead of creating all objects we use during the run method in that
method, we create them now in SharkRunner's constructor, which
semantically also makes more sense. This in turn simplifies the
separation of new private methods that need to use some of these
internal objects (like import_trees and create_thread_objects). This
will also make it easier to further sub-divide the logic in run, which
is the main goal of this exercise.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>