nfw_distribution overhauled
The nfw_distribution class (previously known as NfwDistribution) now
lives in its own header file, so it can be easily imported from more
modules without pulling the entire dark_matter_halos header file.

nfw_distribution also follows RandomNumberDistribution.
RandomNumberDistribution is a C++11 concept with a number of
requirements. Most of these we don't actually need (we actually only
need the () operator), but having them all makes it easier for other
people to reuse this code, or even put it somewhere standing on its own
right.

This improved implementation has also two main advantages over the
original: first, it is templated on the floating point type that it
uses, so it can be used with types other than double (although we
internally still use only the double type). Secondly, and more
importantly, it doesn't internally use the gsl_sf_lambert_W0 function
anymore, therefore dropping this requirement on potential users.
Instead, users need to specialize the lambert_w0 struct for the required
floating point type, and provide an () operator that returns the value
of the Lambert W0 function for a given input. Later in the code we then
specialize the lambert_w0 struct for doubles using the gsl function.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>