Calculation of stellar mass in bursts due to mergers/disk instabilities
- we now follow separately the stellar mass and metals that were formed
due to starbursts triggered by mergers and disk instabilities. This is
done by giving an input to the physical model telling whether this is a
starburst triggered by one or the other process. 

- In galaxy_writer we now write the two properties separately as well.

- There was a bug with the redshift_of_merger calculation in that I was
giving it the snapshot as input rather than the redshift. Now that is
fixed.

- Names of stellar feedback models have been changed to reflect the
paper in which they were presented.