Travis-enabled shark
The travis builds we are setting up here require gcc 5 or greater. This
is because gcc did not fully implement some of the standard library (in
particular the move operations for the iostream classes) until gcc5 and
the corresponding libstdc++. Due to the same reason, when compiling
against clang we install g++-7 (compiling against clang's own
headers/libc++ is otherwise a bit difficult).

On the MacOS side of things we don't have many surprises.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>