Keeping track of galaxy Id and descedant Id, and mean stellar age.
- Several changes included to keep track of galaxy ids and their
descendant id. This involved defining the Id when creating galaxies for
the first time and defining the descendant IDs depending on whether
galaxies will merge at two consecutive snapshots in galaxy mergers (for
type 2), or otherwise their descendantid = galaxyid.

This was also included in the outputs to be written in galaxies_writer.

- Keeping track of the mean stellar age of galaxies by keeping the
accumulated stellar mass formed * age, and the total stellar mass ever
formed (without taking into account stellar population evolution). 

This has now been included in the outputs to be written in
galaxies_writer.

- Outputs were checked and it looks likes the calculations are all
correct.