We require cmake >= 3.2
FindGSL is not present in cmake < 3.2. In such cases cmake default to
using pkg-config to find the library we are looking for, which at least
in the case of GSL does not work correctly when GSL is not in a standard
location.

A way of circumventing that problem is to add the LIBRARY_DIRs with
link_directories, but I am not so sure whether the pkg-config-based
lookup offers these locations to the caller via a GSL_LIBRARY_DIRS
variable or similar. We can study this in the future, but for the time
being cmake 3.2 is not a very hard requirement to meet.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>