Enforcing naming conventions in hdf5::Writer
Now the HDF5 writer class enforces specific naming conventions for the
names of attributes, groups and datasets. Each of these can use a
different naming convention, although all of them default to sneak_case.
A number of unit tests have been added to make sure this is working as
expected.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>