Changing HDF5 names to be snake_case
This includes group, dataset and attribute names. As future work we
still need to use a common convention for acronym-able names, which are
currently running a bit wild.

The standard plotting scripts have also been updated to reflect these
new names. Because of this, old files cannot be read with the new
scripts. If necessary we could support that, but I highly doubt it the
need wil arise.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>