Improved subvolume handling in plot scripts
We now accept space- and comma- separated subvolume specifications. This
will make it easier to call the script from places where the subvolume
list is already available space-separated. On top of that we are now
also more unique-aware, so even if a subvolume is specified twice it
will not be processed twice.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>