Creating standard plots in HPC environments
This commit adds the initial support for requesting the shark-submit
script to produce all standard plots after the shark instances have
finished their execution. The plots are produced only after a
successful execution of all shark instances, and the resulting files
containing the plots are copied into a plots/ subdirectory for easy
access.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>