Transmit location of hpc scripts
The python plotting scripts live (currently) only as part of the git
repository, and therefore we will need the location of the git repo (or
something within) to be able to correctly refer to them later.

The location of shark-run script is of no use, as it is usually copied
over into a different directory at submission time. Instead of this we
use the location of the shark-submit script as a way to establish where
things can be found. To pass down the information from shark-submit to
shark-run we use the SHARK_HPC_DIR environment variable.

This does not take into account the case when shark-submit is called
from the PATH (because we install it via cmake), but this is probably
not going to happen any time soon anyway.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>