Allow arbitrary python executables to be used
Python is used to run the standard plot scripts. Using a non-default
python executable is useful when the python executable available in the
PATH (either because it's installed on a system-level, or as a module)
is not enough to run the scripts, and therefore the python binary from a
virtual environment containing all dependencies must be used instead.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>