Force using the MPL agg backend in shark-run
We know that shark-run runs most likely as part of a queueing
submission, and therefore lacks a graphical interface to show any plots.
To be on the extremely safe side of things we thus use the agg
matplotlib backend when producing the standard plots within shark-run,
as it requires no graphical interface, and exists in every matplotlib
installation.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>