Initial RTD documentation
Instead of keep growing in the README file, we can afford being a bit
more verbose and write more thorough documentation in the form of .rst
files that can be published to RTD.

This initial commit includes documentation on how to build and run shark
in personal computers and HPC systems, and how configuration files and
options look like. Further information should be added regarding
supported models, configuration option names, the format of the output
HDF5 files, and more.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>